Move implementation to the correct modules
