+++ src/lib.rs
@@ -26,16 +26,18 @@ pub mod arena;
 pub mod differentiable;
 pub mod dummy;
 pub mod errors;
+pub mod forward;
+pub mod inverse;
 pub mod mannequin;
-pub mod rigid;
 
 pub use arena::{
     BreadthFirstIterable, DepthFirstArenaTree, DepthFirstIterable, DirectedArenaTree, DirectionIterable, Nodelike,
 };
 pub use differentiable::{Differentiable, VecJacobian};
 pub use errors::MannequinError;
-pub use mannequin::{Forward, Inverse, Mannequin};
-pub use rigid::{accumulate, Rigid, TransformationAccumulation};
+pub use forward::{accumulate, Forward, TransformationAccumulation};
+pub use inverse::Inverse;
+pub use mannequin::{Mannequin, Rigid};
 // Backends
 #[cfg(feature = "ndarray")]
 pub mod ndarray;
