+++ src/mannequin.rs
@@ -4,44 +4,67 @@
  * realistic joints, muscle simulation, or classical inverse kinematics and obstacle avoidance.
  */
 
-use crate::{DepthFirstIterable, Rigid};
-use std::marker::PhantomData;
+use crate::{DepthFirstIterable, Forward, Inverse};
+use std::{fmt::Debug, hash::Hash, marker::PhantomData};
 
-/// Trait representing a stateful forward kinematics algoritm. For instance, it can represent a rigid body mannequin
-/// (e.g., a robot) or softbody/skinning for character animation.
-pub trait Forward<IT, RB>
-where
-    IT: DepthFirstIterable<RB, RB::NodeId>,
-    RB: Rigid,
-{
-    // TODO maybe change to slice
-    type Parameter: IntoIterator<Item = RB::Parameter>;
-    type Transformation;
+/// A Rigid Body represents a single, rigid link connected to other links via a joint.
+/// Synonyms: Bone
+///
+/// Wraps all linear algebra transformations such that backends
+/// only need to implement this trait.
+pub trait Rigid: PartialEq {
+    /// E.g., 4x4 matrix, (3x1, 3x3), quaternions ...
+    type Transformation: Clone + Debug;
+    /// Vec, \[f64;4\], ...
+    type Point;
+    /// typically joint positions (angles/extension), f64, \[f64,3\]
+    type Parameter;
 
-    // TODO this can have a default implementation
-    fn solve(&mut self, tree: &IT, params: Self::Parameter, target_refs: &[RB::NodeId]) -> Vec<Self::Transformation>;
-}
+    // TODO Explain why this is defined on Rigid (the node) and not Mannequin (the tree) .. in short: otherwise this would be another generic and mannequin.rs would be unreadable because of trait bounds. This way it is quite elegant
+    type NodeId: Eq + Hash + Clone + Debug;
 
-/// Trait representing a stateful inverse kinematics algoritm.
-pub trait Inverse<IT, RB, FK>
-where
-    // Avoid mixing of backends
-    IT: DepthFirstIterable<RB, RB::NodeId>,
+    /// Get the Transformation from the parent taking the connecting joint into account
+    fn transform(&self, params: &Self::Parameter) -> Self::Transformation;
 
-    RB: Rigid,
-    FK: Forward<IT, RB, Transformation = Self::Array>,
-{
-    type Parameter: IntoIterator<Item = RB::Parameter>;
-    type Array;
+    /// Transform a point into the world coordinate system
+    fn globalize(&self, other: &Self::Point) -> Self::Point;
 
-    fn solve(
-        &mut self,
-        tree: &IT,
-        fk: &FK,
-        param: Self::Parameter,
-        target_refs: &[RB::NodeId],
-        target_val: &[Self::Array],
-    ) -> Self::Parameter;
+    /// Transform a point into the local coordinate system
+    fn localize(&self, other: &Self::Point) -> Self::Point;
+
+    /// Dimensionality of the partial derivatives (e.g., 3 for position, 6 for position and orientation)
+    fn dim(&self) -> usize;
+
+    /// Compute partial derivative of all effectors
+    /// pose: This node's FoR in global coordinates
+    /// joint: Reference to the joint node
+    /// joint_pose: the joint's FoR in global coordinates
+    /// target_buffer: Memory location to where the results are written too.
+    fn partial_derivative(
+        &self,
+        pose: &Self::Transformation,
+        joint: &Self,
+        joint_pose: &Self::Transformation,
+        target_buffer: &mut [f64],
+    );
+
+    /// number of effectors
+    fn effector_count(&self) -> usize;
+
+    /// The number of rows / elements the effector take in the jacobian matrix (usually dim * cound).
+    /// However, by manually granting contol, one can have effocters with different dimensionality
+    fn effector_size(&self) -> usize {
+        self.dim() * self.effector_count()
+    }
+
+    /// Returns the eutral element wrt. the transoformation convention used
+    fn neutral_element() -> Self::Transformation;
+
+    // Invert a transformation
+    fn invert(trafo: &Self::Transformation) -> Self::Transformation;
+
+    /// Concat two transformations
+    fn concat(first: &Self::Transformation, second: &Self::Transformation) -> Self::Transformation;
 }
 
 /// Struct for holding the composition of character animation algorithms in a flat architecture for
